Compilation fix for companion

+++ drivers.cpp
@@ -258,7 +258,7 @@ void per10ms()
 #if defined(AUDIO)
     uint8_t warble = 0;
 #endif
-    int8_t verticalSpeed = limit((int16_t)-100, frskyHubData.varioSpeed/10, (int16_t)+100);
+    int8_t verticalSpeed = limit((int16_t)-100, (int16_t)(frskyHubData.varioSpeed/10), (int16_t)+100);
 
     uint16_t interval;
     if (verticalSpeed == 0) {
